backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3
psycopg2-binary==2.9.9
dj-database-url==2.1.0
pandas==2.1.1
openpyxl==3.1.2
python-dotenv==1.0.0

